Remove the RamAvailable component of ram module
