+++ src/routes/[network]/(account)/ram/buy/+page.svelte
@@ -17,11 +17,10 @@
 	import BytesInput from '$lib/components/input/bytes.svelte';
 	import Card from '$lib/components/layout/box/card.svelte';
 	import AssetText from '$lib/components/elements/asset.svelte';
+	import RamResource from '$lib/components/elements/ramresource.svelte';
 
 	import { BuyRAMState } from './state.svelte';
-	import { preventDefault } from '$lib/utils';
-
-	import AvailableRam from '../(components)/availableRam.svelte';
+	import { calAvailableSize, preventDefault } from '$lib/utils';
 
 	let bytesInput: BytesInput | undefined = $state();
 	let assetInput: AssetInput | undefined = $state();
@@ -31,6 +30,7 @@
 	const debugMode = getSetting('debug-mode', true);
 
 	const buyRamState: BuyRAMState = $state(new BuyRAMState(data.network.chain));
+	const ramAvailableSize = $derived(calAvailableSize(context.account?.ram));
 
 	let transactionId: Checksum256 | undefined = $state();
 
@@ -99,7 +99,7 @@
 
 <Card>
 	<form onsubmit={preventDefault(handleBuyRAM)} class="mx-auto max-w-2xl space-y-4">
-		<AvailableRam />
+		<RamResource ramAvailable={ramAvailableSize} />
 
 		<Stack class="gap-3">
 			<Label class="text-lg" for="bytesInput">Amount to buy</Label>
